♻️ 댓글 영역 > 댓글 컴포넌트 분리를 통해 더보기 버튼 클릭 시 모든 댓글의 모달이 나타나는 문제 해결
